Set TTL on BuildCache items for travis.
This PR makes all writes from TravisBuildMonitor to BuildCache
include a TTL. The TTL defaults to 60 days and can be overridden
by the property travis.cachedJobTTLDays.

setBuildCacheTTL() is called every time TravisBuildMonitor starts
(igor starts) that checks is all jobs in the BuildCache has ttl
set. If not, it will set it.