Travis - make igor able to override multiple properties in build
The travis api acts strange when we trigger builds. We need to inject one property in `env` to be able to override `global_env`. We have to override `global_env` to override multiple properties in one job, if we do that in `env` it will trigger a job in travis for every property as a build matrix. The data inside `config.global_env` can both include strings and secrets `"MYPROPERTY=foo"` and secrets in a map like `{"secure":"xyz"}`. The code in this PR gathers the secrets from the last build and injects them into the payload when we trigger builds with igor against travis.