Travis include artifacts in trigger event.
We want to implement support for multirepo incremental builds in spinnaker.
Given a project that creates two or more services and that is building incrementally,
we want to be able to filter in the trigger stage on which artifacts that have been uploaded.

This will let us have many pipelines listening to the same build, but only trigger them
when the artifact we will use in the bake is present.