fix(jenkins): Use new location for revision information (#367) (#376)
The jenkins-git plugin is removing the BuildData structure
from the results of the build as of version 4.0.0 of the
plugin. Spinnaker uses that data to populate the 'scm'
information for the build, so users who have upgraded
their jenkins-git plugin are not getting any SCM information
in Spinnaker.

The replacement for BuildData is a new structure BuildDetails.
All of the information we were fetching from BuildData is
available in BuildDetails.

Update igor to look for the SCM information in BuildDetails if
nothing is found in BuildData. This change is backwards-compatible
as the new BuildDetails location is only queried if there is
nothing in the old BuildData location.