feat(travis): Customize the Travis build message on triggered builds
Lets you customize the build message used when Spinnaker triggers builds in Travis. If you set a custom parameter in the Travis stage in Spinnaker with the value of this property as the key (e.g `travis.buildMessage=My customized message`, the build message in Travis will be _Triggered by Spinnaker: My customized message_. The first part of this message is not customizable.