fix(travis): Check that logs are complete before parsing them
The Travis™ API often returns incomplete logs for a few moments (~30-60 seconds) after a build completes. As we are dependent on the logs for detecting both build artifacts and fetching properties, I've added additional checks that will cause Spinnaker to wait until the logs are ready before trying to parse them. If you thought that there's an endpoint in the Travis API to check if the logs are ready yet, you would be wrong. So we have to check this by actually downloading the logs multiple times until they are complete.

**Pros:** We get rid of a whole bunch of weird, non-deterministic bugs caused by the last few lines of the log to sometimes be cut away. Especially since the last couple of lines is where users usually upload artifacts, our pipelines will now Work More Often™ and Fail Less™ because of missing artifacts and/or properties.

**Cons:** It will take, on average, 30 seconds longer before a completed build triggers a Spinnaker pipeline or a Travis stage completes. And we're probably causing some increased load on Travis since we are asking for logs more often.

**Bonus:** I've added Redis caching for the logs as well, with a TTL of 10 minutes. As soon as the log is ready, it goes into the cache.