fix(gcb): Fix deserialization of Build object
Jackson's ObjectMapper does not properly deserialize Build objects,
as nested properties that extend com.google.api.client.json.GenericJson
(which itself implements Map) are are deserialized as Maps rather than
as the specific type.

There is probably a way of configuring ObjectMapper to do this properly
but I have not (yet) been able to find it. As a workaround, use the
Google Cloud supported deserialization classes to deserialize this
JSON. In order to keep this workaround localized, make a small
class that will handle deserializing Build classes.

This commit also adds BuildOptions to the build in the integration
test; adding this causes the test to fail when using the standard
ObjectMapper for deserialization.