feat(gcb): Add endpoints for storing and retrieving build status
Create and endpoint to accept the current status of a build and store
it in Redis.  This endpoint is intended to be called by echo when it
receives a PubSub message about a build status change.  We take care
to look at the timestamp of incoming messages and only update the
cache if the incoming information is newer, as PubSub does not guarantee
delivery order.

Create an endpoint to return the current status of the build, as
per the cache.  This commit does not implement any fallback to
calling the API if the cache does not contain the build; this
will be added in a future commit.