feat(gcb): Create build should return only the build
When creating a build, we should return only the build instead of the
operation. In general, callers will only be interested in the id of the
created build, and it will be easier for callers to just look up the
"id" of the returned object instead of looking at the deeply nested
operation -> metadata -> build -> id field.