refactor(gcb): Extract all interactions with GCB API to a client class
Currently GoogleCloudBuildAccount directly generates calls to the GCB
API (though uses a GoogleCloudBuildExecutor to actually execute the
requests). As we add more complex logic to GoogleCloudBuildAccount and
want to add some unit tests, it makes sense to pull the logic that is
just generating API calls into a separate class so it can be easily
stubbed.

Create a GoogleCloudBuildClient which will expose GCB API operations
and execute them for the consumer.