feat(gcb): Add ttl to build cache entries and fall back to polling
In order to handle cases where a pubsub message is not delivered,
we'd like to fall back to polling if we haven't received an update
on a build in some amount of time.

Configure this by adding a TTL on all entries in the cache; we'll
set a shorter TTL for in-progress builds (as we expect an update)
and a longer one for completed builds (which should never have
another status change). If we check the cache and there is no build
there, fall back to an external poll, and store that result for
the same amount of time.