feat(gcb): Add GCS object to GCB account
As Google Cloud Build stores a manifest of produced artifacts in
Google Cloud Storage, a GCB account will need to be able to talk
to GCS in order to return information about produced artifacts.

This allows us to better encapsulate information about produced artifacts
rather than having the GCB account simply return a path to the manfest
and require the caller to parse any artifacts (which are in a GCB-specific
format).