feat(gcb): Add endpoint to fetch build artifacts
The endpoint should return a list of Spinnaker artifacts representing
artifacts produced by the build.  A build can produce either docker
containers or arbitrary files in Google Cloud Storage.

Docker containers are directly included in the build output, so we
can read the relevant fields and construct an artifact. GCS files
are not included directly in the output; the output contains a
GCS path to a manifest of produced files. We need to fetch and
parse this manifest in order to generate the produced artifacts.