fix(artifactory): correctly construct Artifactory URL
Previously we got the base URL from the client which unintuitively removed the context root for us. Now we grab it from the search which has the artifactory base URLs that Igor is configured with.
We also provide a more specific URL that identifies a particular pom, rather than just the directory of the group, artifact, and version.