fix(gcb): Correctly fetch artifact manifest when it contains a version
The logic to fetch an artifact manifest does not properly account for
the GCS path containing a version (as in gs://bucket/file.json#123).

This isn't causing any active issues as GCB creates a file named
with a UUID and does not include the version in the build output,
but we should properly handle this in case this changes (or in
case we re-use the artifact fetching code elsewhere).