chore(health): delete PollingMonitorHealth
It is currently a complicated no-op health check, it always returns true.
The reason is that upOnce starts false and can never be set to true,
so the entire logic is bypassed... which I think is a good thing, I don't
think we would want igor to become unhealthy just because one or more
poller is having issues. It could cause igor to knock itself out of a load
balancer when it could in fact still be serving requests, and this would
cause pipeline stages to fail.