refactor(gcb): Make packages package-private and final where possible
The only external contract the com.netflix.spinnaker.igor.gcb
package intends to provide is the GoogleCloudBuildController
component; everything else is an implementation detail.

To better communicate and enforce this:
* Make the controller (and all mapped methods) public; some of
these were package-private before which works because Spring uses
reflection to call controllers, but we should document them
accurately as being called outside of the package.
* Make all other classes (and methods) that were public package-private

Finally (no pun intended), none of the classes in this package
were designed for inheritance so mark them final wherever possible.
The cases where it is not possible are:
* Classes that are mocked in tests as a final class cannot be
mocked; in this case, we'll rely on the fact that the class is
package-private so at least can't be subclassed outside the package.
* Spring controller and configuration classes aren't allowed to be
final