refactor(gcb): Add nullity annotations
Add default Nonnull to all methods, parameters and fields across the
package, and annotate the few exceptions with Nullable.

This does surface a few places where we could consider replacing a
null with either an Optional or an appropriate null object, but
this PR does not tackle any of those refactors.