refactor(gcb): Make GCB account repository immutable
As we never add/remove accounts after startup, the repository
should be immutable. Maybe at some point if we support dynamic
accounts we should change this (but at that point we'd need to
think about thread safety anyway).

This commit also makes a few changes to the tests on that class;
they had to be updated anyway to use the new Builder class, and
while I was there I improved the assertions to make them more clear.