test(gcb): Update tests to run postFilter logic
The GCB integration tests currently don't run the postFilter logic
that is added as an annotation on controller methods. I believe this
is because we've been to selective in exactly the beans that should
be present in the test, and are not pulling in the required Spring
Security beans.

In order to fix this, add the same @ComponentScan that we have on
Main.class to the test so it pulls in the same beans.  (We can then
also remove the specific controller bean we'd been pulling in, but
will leave all the configuration beans in place.)

This causes the listAccountTest to fail because the listAccounts
function is currently broken; the next commit will fix the function
and the test.