fix(gcb): Return mutable lists from methods annoated with PostFilter
Controller methods annotated with PostFilter cannot return immutable
lists because the filtering is done in-place on the returned array;
update all controller methods in GoogleCloudBuildController to return
a mutable list.

As we would like the lower levels of the stack to deal in immutable
collections as much as possible, just create a mutable collection
at the last step before returning from the controller method.