fix(gcb): Use mock HystrixSpectatorPublisher bean
The tests were failing because we can only ever create a single
HystrixSpectatorPublisher bean, due to the fact that it mutates
a static field on the HystrixPlugins class, which causes a
failure if we try to mutate it again.

This was addressed in MainTest by clearing the Hystrix class after
every test, but let's just inject a mock bean instead to avoid
mutating the static class in the first place.