fix(decorator): Expect at least 3 fields in debian package name
When a debian package name has no underscores in its name, the method
extractVersion throws the exception java.util.NoSuchElementException because it
cannot pop an empty list. When it happens Igor stops processing the repos
(Travis in my case) and the triggers.

This fix ensures the debian package name is composed of at least 3 fields in
order to enforce the naming convention <name>_<version>_<architecture>.deb