config(core): Push Halyard config to base
The redis url is being set in the Halyard-specific file that is
concatenated to the config; let's just move that defaulting to
the base igor.yml.

This should not cause any visible change to non-Halyard users, as
we are defaulting back to the same localhost:6379 if the redis
override is not defined. It will also not affect Halyard users as
it's just changing where we set this value for them.