refactor(gcb): Make GCB account immutable and null-safe
This commit moves the GoogleCloudBuildProperties.Account class
to be immutable and use constructor binding. It also makes the
minor change of storing the result of Permissions.build() instead
of the raw builder, rather than calling Permissions.build() every
time we ask for the permissions.

It also annotates the class as NonnullByDefault, and handles any
necessary null defaulting in the constructor.