fix(jenkins): Don't kick off multiples of the same build
The issue here is that when an operation takes a long time (e.g. talking to jenkins to kick off a build)
this prolonged time can cause orca to try to resubmit the request multiple times thus kicking off multiple (or infinite number of) builds.

We solve this, but setting a flag that a given request (keyed of X-SPINNAKER-REQUEST-ID and master/job/param) is being processed so if an identical
request comes in the controller can detect that and respond with 202 ACCEPTED. This will cause orca to retry (corresponding PR coming).
When the job is finally kicked off, the resulting build number will be stored in the cache and next time orca queries it will get the build number back