refactor(igor): Rename master to controller
This change renames master to controller. The reasoning behind this
change is that none of the services seems to be using that terminology
anymore.

Changing the terminology in Igor to something that at least some of
the services uses makes it easier to understand what's being referenced.

All externally accessible APIs with the old terminology are marked as
deprecated and will use the new functions until the deprecated functions
are finally removed. This to remain backward compatible.