fix(travis): Redis TTL is given in seconds, not minutes
This fixes a rare bug where some builds will "be lost" and not trigger any pipelines. Because we effectively set the TTL to 5 minutes instead of 5 hours, Spinnaker would quickly stop tracking builds. Usually this is no problem because they will be present in the builds returned from the Travis API, but if the `numberOfJobs` property is set too low or during periods with a lot of simultaneous builds, this could cause builds (especially long running ones) to be lost.