fix(travis): Don't panic if log fetching fails
If a user has deleted the log for a job in Travis or it has expired due too old age, Spinnaker will fail to display the list of builds in the start manual execution box, even though there are other builds/jobs available. This fix will just filter these (and other) failures, and instead of failing completely, it will print an error message in the Igor log.
The `catch` clause will print a log line at level "INFO" for cases where the log has expired (which can happen from time to time) and at WARN level for other (actual) errors.